% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{hchart_corr}
\alias{hchart_corr}
\title{Highchart Correlation Plot}
\usage{
hchart_corr(corr_mat)
}
\arguments{
\item{corr_mat}{an matrix, usually returned by \link[stats:cor]{stats::cor}}
}
\description{
Highchart Correlation Plot
}
