% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{predizer_desligamento}
\alias{predizer_desligamento}
\title{Faz predição de desligamento}
\usage{
predizer_desligamento(novo_funcionario)
}
\arguments{
\item{novo_funcionario}{data.frame com as informacoes do funcionario}
}
\description{
Faz predição de desligamento
}
